Moved fetching list of tools to tool service for both admin and employees, resolved promise on checkAdmin, Added additional attributes for UserTools model

+++ server/services/tool/tool.js
@@ -17,9 +17,11 @@ function add(request, response) {
       .save()
       .then(doc => {
         new UserTools({
-          toolName: request.body.name.charAt(0).toUpperCase() + request.body.name.slice(1),
-          toolID: doc._id,
-          userID: request.body.userID
+          name: request.body.name.charAt(0).toUpperCase() + request.body.name.slice(1),
+          homePage: request.body.link,
+          userID: request.body.userID,
+          logoPath: logoPath,
+          className: request.body.name.toLowerCase(),
         })
         .save()
         .then(doc => {
@@ -50,9 +52,11 @@ function add(request, response) {
       })
     } else {
       new UserTools({
-        toolName: request.body.name.charAt(0).toUpperCase() + request.body.name.slice(1),
-        toolID: tool._id,
-        userID: request.body.userID
+        name: request.body.name.charAt(0).toUpperCase() + request.body.name.slice(1),
+        homePage: request.body.link,
+        userID: request.body.userID,
+        logoPath: logoPath,
+        className: request.body.name.toLowerCase(),
       })
       .save()
       .then(doc => {
@@ -76,17 +80,32 @@ function add(request, response) {
 }
 
 function get(request, response) {
-  Tools.find({})
+  const {
+    role, id,
+  } = request.query;
+
+  if (role !== undefined && role === 'Admin' && role.toLowerCase() === 'admin') {
+    Tools.find({})
+    .sort({'createdAt': 'desc'})
     .then(docs => {
       return response.json(docs);
     })
     .catch(error => {
       return response.json({ success: false, message: error.ValidationError });
     });
+  } else {
+    UserTools
+    .find({ userID: request.query.id })
+    .sort({'createdAt': 'desc'})
+    .exec((error, employeeTools) => {
+      if (error) return response.json(error);
+      return response.json(employeeTools);
+    });
+  }
 }
 
 function remove(request, response) {
-  if (request.body.admin.role === 'Admin') {
+  if (request.body.admin.role === 'Admin' && request.body.admin.role.toLowerCase() === 'admin') {
     Tools
     .findOneAndDelete({ _id: request.body.toolID })
     .exec((error, doc) => {
@@ -103,6 +122,24 @@ function remove(request, response) {
       .catch(error => {
         return response.json({success: false, message: error});
       })
+    });
+  } else {
+    UserTools
+    .findOneAndDelete({ _id: request.body.toolID })
+    .exec((error, doc) => {
+      if (error) return response.json(error);
+      UserTools
+        .find({})
+        .then(docs => {
+          response.json({
+            'success': true,
+            'message': `${doc.name} is deleted successfully.`,
+            'tools': docs
+          });
+        })
+        .catch(error => {
+          return response.json({success: false, message: error});
+        })
     });
   }
 }
